Major codebase improvements: optimized solvers, trajectory classification fixes, and LGST example files
- Added optimized solver implementations (SolverSharedCodePlusSolar_Optimized, StochasticInputRK45Solver_Vectorized)
- Fixed integer overflow bug in TrajectoryClassification.py and TrajectoryClassification_numpy.py
- Added three LGST simulation example files with different optimization levels
- Updated imports across solver files to use optimized versions
- Added comprehensive documentation (QUICK_START_GUIDE, OPTIMIZATION_COMPARISON, DIRECTORY_STRUCTURE_README)
- Cleaned up old test results and debug files
- Updated .gitignore to exclude simulation results directories

+++ TrajectoryClassification_numpy.py
@@ -60,9 +60,9 @@ def _process_trajectory_numpy(trajectories: np.ndarray, alpha: float, beta: floa
     """
     # Convert to numpy array if it's a DataFrame
     if hasattr(trajectories, 'values'):
-        traj = trajectories.values[:, :3]  # Take first 3 columns if more exist
+        traj = trajectories.values[:, :3].astype(np.float64)  # Force float64 to avoid overflow
     else:
-        traj = np.asarray(trajectories)[:, :3]  # Ensure numpy array
+        traj = np.asarray(trajectories, dtype=np.float64)[:, :3]  # Ensure float64 array
     
     # Pre-compute all positions and distances for better performance
     x = traj[:, 0]
